Edit utils.py modify_change_name_extension, delete unrelated functions.
Edit utils.py modify change_name_extension.

Edit utils.py clean up and delete functions in utils.
