+++ __init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
