+++ copy_and_align.py
@@ -31,7 +31,7 @@ midi_files = [el.strip() for el in lines]
 '''
 
 # read from folder
-midi_files = utils.find_files_in_subdirs(INPUT_DIR, '*.mid')
+midi_files = utils.find_files_in_subdir(INPUT_DIR, '*.mid')
 
 n_match = 0
 n_unmatch = 0
@@ -46,7 +46,7 @@ for midi_file in midi_files:
         n_match += 1
         continue
 
-    file_folder, file_name = utils.split_path_from_path(midi_file)
+    file_folder, file_name = utils.split_head_and_tail(midi_file)
     perform_midi = midi_file
     score_midi = os.path.join(file_folder, 'midi.mid')
     print perform_midi
